backend/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.0.10
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.14.1
drf-yasg==1.21.7
flake8==7.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.9
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.10.1
pytest==8.0.2
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
stripe==7.14.0
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
